First refactoring of the assembly.
